Consolidate CLF-CBF example script
