+++ .gitignore
@@ -148,9 +148,7 @@ site/
 
 # camera params saved during rendering
 DepthCamera_*.json
-DepthCapture_*.pngforce_goal_distance_plot.pdf
-force_goal_distance_plot.png
-videos/
+DepthCapture_*.png
 
 # Generated CLF-CBF outputs
 force_goal_distance_plot.pdf
